+++ src/error.rs
@@ -3,12 +3,15 @@
 pub use actix_http::error::*;
 use derive_more::{Display, Error, From};
 use serde_json::error::Error as JsonError;
+use serde_urlencoded::de::Error as FormDeError;
+use serde_urlencoded::ser::Error as FormError;
 use url::ParseError as UrlParseError;
 
-use crate::{http::StatusCode, HttpResponse};
+use crate::http::StatusCode;
 
 /// Errors which can occur when attempting to generate resource uri.
-#[derive(Debug, PartialEq, Display, From)]
+#[derive(Debug, PartialEq, Display, Error, From)]
+#[non_exhaustive]
 pub enum UrlGenerationError {
     /// Resource not found
     #[display(fmt = "Resource not found")]
@@ -23,13 +26,12 @@ pub enum UrlGenerationError {
     ParseError(UrlParseError),
 }
 
-impl std::error::Error for UrlGenerationError {}
-
 /// `InternalServerError` for `UrlGeneratorError`
 impl ResponseError for UrlGenerationError {}
 
 /// A set of errors that can occur during parsing urlencoded payloads
 #[derive(Debug, Display, Error, From)]
+#[non_exhaustive]
 pub enum UrlencodedError {
     /// Can not decode chunked transfer encoding.
     #[display(fmt = "Can not decode chunked transfer encoding.")]
@@ -52,8 +54,16 @@ pub enum UrlencodedError {
     ContentType,
 
     /// Parse error.
-    #[display(fmt = "Parse error.")]
-    Parse,
+    #[display(fmt = "Parse error: {}.", _0)]
+    Parse(FormDeError),
+
+    /// Encoding error.
+    #[display(fmt = "Encoding error.")]
+    Encoding,
+
+    /// Serialize error.
+    #[display(fmt = "Serialize error: {}.", _0)]
+    Serialize(FormError),
 
     /// Payload error.
     #[display(fmt = "Error that occur during reading payload: {}.", _0)]
@@ -63,52 +73,65 @@ pub enum UrlencodedError {
 /// Return `BadRequest` for `UrlencodedError`
 impl ResponseError for UrlencodedError {
     fn status_code(&self) -> StatusCode {
-        match *self {
-            UrlencodedError::Overflow { .. } => StatusCode::PAYLOAD_TOO_LARGE,
-            UrlencodedError::UnknownLength => StatusCode::LENGTH_REQUIRED,
+        match self {
+            Self::Overflow { .. } => StatusCode::PAYLOAD_TOO_LARGE,
+            Self::UnknownLength => StatusCode::LENGTH_REQUIRED,
+            Self::Payload(err) => err.status_code(),
             _ => StatusCode::BAD_REQUEST,
         }
     }
 }
 
 /// A set of errors that can occur during parsing json payloads
-#[derive(Debug, Display, From)]
+#[derive(Debug, Display, Error)]
+#[non_exhaustive]
 pub enum JsonPayloadError {
     /// Payload size is bigger than allowed. (default: 32kB)
     #[display(fmt = "Json payload size is bigger than allowed")]
     Overflow,
+
     /// Content type error
     #[display(fmt = "Content type error")]
     ContentType,
+
     /// Deserialize error
     #[display(fmt = "Json deserialize error: {}", _0)]
     Deserialize(JsonError),
+
+    /// Serialize error
+    #[display(fmt = "Json serialize error: {}", _0)]
+    Serialize(JsonError),
+
     /// Payload error
     #[display(fmt = "Error that occur during reading payload: {}", _0)]
     Payload(PayloadError),
 }
 
-impl std::error::Error for JsonPayloadError {}
+impl From<PayloadError> for JsonPayloadError {
+    fn from(err: PayloadError) -> Self {
+        Self::Payload(err)
+    }
+}
 
-/// Return `BadRequest` for `JsonPayloadError`
 impl ResponseError for JsonPayloadError {
-    fn error_response(&self) -> HttpResponse {
-        match *self {
-            JsonPayloadError::Overflow => HttpResponse::new(StatusCode::PAYLOAD_TOO_LARGE),
-            _ => HttpResponse::new(StatusCode::BAD_REQUEST),
+    fn status_code(&self) -> StatusCode {
+        match self {
+            Self::Overflow => StatusCode::PAYLOAD_TOO_LARGE,
+            Self::Serialize(_) => StatusCode::INTERNAL_SERVER_ERROR,
+            Self::Payload(err) => err.status_code(),
+            _ => StatusCode::BAD_REQUEST,
         }
     }
 }
 
 /// A set of errors that can occur during parsing request paths
-#[derive(Debug, Display, From)]
+#[derive(Debug, Display, Error)]
+#[non_exhaustive]
 pub enum PathError {
     /// Deserialize error
     #[display(fmt = "Path deserialize error: {}", _0)]
     Deserialize(serde::de::value::Error),
 }
-
-impl std::error::Error for PathError {}
 
 /// Return `BadRequest` for `PathError`
 impl ResponseError for PathError {
@@ -119,6 +142,7 @@ impl ResponseError for PathError {
 
 /// A set of errors that can occur during parsing query strings.
 #[derive(Debug, Display, Error, From)]
+#[non_exhaustive]
 pub enum QueryPayloadError {
     /// Query deserialize error.
     #[display(fmt = "Query deserialize error: {}", _0)]
@@ -133,24 +157,25 @@ impl ResponseError for QueryPayloadError {
 }
 
 /// Error type returned when reading body as lines.
-#[derive(From, Display, Debug)]
+#[derive(Debug, Display, Error, From)]
+#[non_exhaustive]
 pub enum ReadlinesError {
-    /// Error when decoding a line.
     #[display(fmt = "Encoding error")]
     /// Payload size is bigger than allowed. (default: 256kB)
     EncodingError,
+
     /// Payload error.
     #[display(fmt = "Error that occur during reading payload: {}", _0)]
     Payload(PayloadError),
+
     /// Line limit exceeded.
     #[display(fmt = "Line limit exceeded")]
     LimitOverflow,
+
     /// ContentType error.
     #[display(fmt = "Content-type error")]
     ContentTypeError(ContentTypeError),
 }
-
-impl std::error::Error for ReadlinesError {}
 
 /// Return `BadRequest` for `ReadlinesError`
 impl ResponseError for ReadlinesError {
@@ -168,26 +193,25 @@ mod tests {
 
     #[test]
     fn test_urlencoded_error() {
-        let resp: HttpResponse =
-            UrlencodedError::Overflow { size: 0, limit: 0 }.error_response();
+        let resp = UrlencodedError::Overflow { size: 0, limit: 0 }.error_response();
         assert_eq!(resp.status(), StatusCode::PAYLOAD_TOO_LARGE);
-        let resp: HttpResponse = UrlencodedError::UnknownLength.error_response();
+        let resp = UrlencodedError::UnknownLength.error_response();
         assert_eq!(resp.status(), StatusCode::LENGTH_REQUIRED);
-        let resp: HttpResponse = UrlencodedError::ContentType.error_response();
+        let resp = UrlencodedError::ContentType.error_response();
         assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
     }
 
     #[test]
     fn test_json_payload_error() {
-        let resp: HttpResponse = JsonPayloadError::Overflow.error_response();
+        let resp = JsonPayloadError::Overflow.error_response();
         assert_eq!(resp.status(), StatusCode::PAYLOAD_TOO_LARGE);
-        let resp: HttpResponse = JsonPayloadError::ContentType.error_response();
+        let resp = JsonPayloadError::ContentType.error_response();
         assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
     }
 
     #[test]
     fn test_query_payload_error() {
-        let resp: HttpResponse = QueryPayloadError::Deserialize(
+        let resp = QueryPayloadError::Deserialize(
             serde_urlencoded::from_str::<i32>("bad query").unwrap_err(),
         )
         .error_response();
@@ -196,9 +220,9 @@ mod tests {
 
     #[test]
     fn test_readlines_error() {
-        let resp: HttpResponse = ReadlinesError::LimitOverflow.error_response();
+        let resp = ReadlinesError::LimitOverflow.error_response();
         assert_eq!(resp.status(), StatusCode::PAYLOAD_TOO_LARGE);
-        let resp: HttpResponse = ReadlinesError::EncodingError.error_response();
+        let resp = ReadlinesError::EncodingError.error_response();
         assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
     }
 }